+++ src/emulates/Util.h
@@ -1,6 +1,7 @@
 #ifndef UTIL_H
 #define UTIL_H
 #include <stdbool.h>
+#include <stdint.h>
 #define MAX_MEMORY 2100000
 struct PSTATE
 {
@@ -9,17 +10,17 @@ struct PSTATE
 
 struct Registers {
     // int R0 to R30 for 31 genreal purpose registers
-    long long general[31];
+    int64_t general[31];
     // Special registers
-    long long PC, ZR, SP;
+    int64_t PC, ZR, SP;
     // processor state register
     struct PSTATE pstate;
 };
 #endif
 
-extern unsigned int getBitRange(int num, int x, int y);
-extern unsigned int get_bit(int startBit, int numBits, unsigned int number);
-extern unsigned long long get_bitl(int startBit, int numBits, unsigned long long number);
+extern uint32_t getBitRange(int num, int x, int y);
+extern uint32_t get_bit(int startBit, int numBits, uint32_t number);
+extern uint64_t get_bitl(int startBit, int numBits, uint64_t number);
 extern int sget_bit(int startBit, int numBits, int number);
-extern bool hasCarryBorrow(long long a, long long b, bool is_64, bool is_sub);
-extern bool overflow (long long a, long long b , bool is_64);
+extern bool hasCarryBorrow(int64_t a, int64_t b, bool is_64, bool is_sub);
+extern bool overflow (int64_t a, int64_t b , bool is_64);
