+++ src/emulates/inputformat.h
@@ -1,25 +1,26 @@
 #include <stdbool.h>
+#include <stdint.h>
 #ifndef INPUTFORMAT_H
 #define INPUTFORMAT_H
 struct send_DPI
 {
     bool sf;
-    int opc;
-    int opi;
-    int operand;
-    int rd;
+    uint8_t opc;
+    uint8_t opi;
+    int32_t operand;
+    uint8_t rd;
 };
 
 struct send_DPR
 {
     bool sf;
     bool M;
-    int opc;
-    int opr;
-    int rm;
-    int operand;
-    int rn;
-    int rd;
+    uint8_t opc;
+    uint8_t opr;
+    uint8_t rm;
+    uint8_t operand;
+    uint8_t rn;
+    uint8_t rd;
 };
 
 struct sdtp
@@ -27,26 +28,26 @@ struct sdtp
     bool sf;
     bool L;
     bool U;
-    int offset;
-    int xn;
-    int rt;
+    int16_t offset;
+    uint8_t xn;
+    uint8_t rt;
 };
 
 struct loadliteral
 {
     bool sf;
-    long long simm19;
-    int rt;
+    int32_t simm19;
+    uint8_t rt;
 };
 
 struct send_branch
 {
-    int sf;
-    int operand;
+    uint8_t sf;
+    int32_t operand;
 };
 #endif
-extern struct send_DPI to_DPI(unsigned int instruction);
-extern struct send_DPR to_DPR(unsigned int instruction);
-extern struct sdtp to_SDT(unsigned int instruction);
-extern struct loadliteral to_LL(unsigned int instruction);
-extern struct send_branch to_branch(unsigned int instruction);
+extern struct send_DPI to_DPI(uint32_t instruction);
+extern struct send_DPR to_DPR(uint32_t instruction);
+extern struct sdtp to_SDT(uint32_t instruction);
+extern struct loadliteral to_LL(uint32_t instruction);
+extern struct send_branch to_branch(uint32_t instruction);
