+++ src/emulates/single_data_transfer/sdt.h
@@ -1,4 +1,4 @@
 #include "../Util.h"
 #include "../inputformat.h"
-extern long long getRegisterValue(int n, struct Registers* regs);
+extern int64_t getRegisterValue(int n, struct Registers* regs);
 extern void SingleDataTransfer(char* memory, struct Registers* regs, struct sdtp s);
